[mypyc] Refactor: extract vtable code from genops
